move tests to separate class
